+++ components/ehmtxv2/EHMTX.h
@@ -50,6 +50,7 @@ namespace esphome
   class EHMTX_queue;
   class EHMTX_Icon;
   class EHMTXNextScreenTrigger;
+  class EHMTXEmptyQueueTrigger;
   class EHMTXAddScreenTrigger;
   class EHMTXIconErrorTrigger;
   class EHMTXExpiredScreenTrigger;
@@ -66,6 +67,7 @@ namespace esphome
     bool show_day_of_week;
 
     std::vector<EHMTXNextScreenTrigger *> on_next_screen_triggers_;
+    std::vector<EHMTXEmptyQueueTrigger *> on_empty_queue_triggers_;
     std::vector<EHMTXIconErrorTrigger *> on_icon_error_triggers_;
     std::vector<EHMTXExpiredScreenTrigger *> on_expired_screen_triggers_;
     std::vector<EHMTXNextClockTrigger *> on_next_clock_triggers_;
@@ -181,6 +183,7 @@ namespace esphome
     void draw_lindicator();
 
     void add_on_next_screen_trigger(EHMTXNextScreenTrigger *t) { this->on_next_screen_triggers_.push_back(t); }
+    void add_on_empty_queue_trigger(EHMTXEmptyQueueTrigger *t) { this->on_empty_queue_triggers_.push_back(t); }
     void add_on_add_screen_trigger(EHMTXAddScreenTrigger *t) { this->on_add_screen_triggers_.push_back(t); }
     void add_on_icon_error_trigger(EHMTXIconErrorTrigger *t) { this->on_icon_error_triggers_.push_back(t); }
     void add_on_expired_screen_trigger(EHMTXExpiredScreenTrigger *t) { this->on_expired_screen_triggers_.push_back(t); }
@@ -239,6 +242,13 @@ namespace esphome
   public:
     explicit EHMTXNextScreenTrigger(EHMTX *parent) { parent->add_on_next_screen_trigger(this); }
     void process(std::string, std::string);
+  };
+
+  class EHMTXEmptyQueueTrigger : public Trigger<>
+  {
+  public:
+    explicit EHMTXEmptyQueueTrigger(EHMTX *parent) { parent->add_on_empty_queue_trigger(this); }
+    void process();
   };
 
   class EHMTXAddScreenTrigger : public Trigger<std::string, uint8_t>